requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
MarkupSafe==3.0.2
pymongo==4.7.0
python-dotenv==1.0.1
python-telegram-bot==20.7
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
tornado==6.3.3
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.3.7
Flask==2.2.2  # Added Flask to handle webhooks
